+++ src/__tests__/common.ts
@@ -4,6 +4,7 @@ import faker from '@faker-js/faker';
 import { mockEnokiFlow } from "./auth";
 import EnokiSigner from "./enoki/signer";
 import { status } from "../constants";
+import { server } from "./server";
 
 export const TESTING_ENV = "testnet";
 
@@ -25,8 +26,12 @@ export async function setupVault(isPublic = false): Promise<string> {
   return vaultId;
 }
 
-export async function cleanup(akord: Akord, vaultId: string): Promise<void> {
-  if (vaultId) {
+export async function cleanup(akord?: Akord, vaultId?: string): Promise<void> {
+  jest.clearAllTimers();
+  if (server) {
+    server.close();
+  }
+  if (akord && vaultId) {
     await akord.vault.delete(vaultId);
   }
 }